refactor(web): fold AdminSidebarShell into CreateConnectorSidebar
`CreateConnectorSidebar` was the only user of the shell. Move the chrome
into it and drop the extra file.

Rename the footer tab to "Exit Connector Setup" with an X icon, to match
"Exit Admin Panel" in `AdminSidebar`.
